Scott - Started working on input capture code
Wrote the initialization routines for all 4 input capture inputs for the
motor encoders. Also setup the input cap interrupt routines in the main
C file. Still need to test initializations and find a way to determine
direction of rotation. Got PCBs in today and both motors now work with
the H-Bridge